+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/sinks/SelectTableSinkBase.java
@@ -20,6 +20,7 @@ package org.apache.flink.table.planner.sinks;
 
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
+import org.apache.flink.core.execution.JobClient;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.datastream.DataStreamSink;
 import org.apache.flink.streaming.api.operators.collect.CollectResultIterator;
@@ -37,6 +38,7 @@ import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.types.Row;
 
+import java.util.Iterator;
 import java.util.UUID;
 import java.util.stream.Stream;
 
@@ -49,12 +51,12 @@ public abstract class SelectTableSinkBase<T> implements StreamTableSink<T> {
 	protected final DataFormatConverters.DataFormatConverter<RowData, Row> converter;
 
 	private final CollectSinkOperatorFactory<T> factory;
-	protected final CollectResultIterator<T> iterator;
+	private final CollectResultIterator<T> iterator;
 
 	@SuppressWarnings("unchecked")
 	public SelectTableSinkBase(TableSchema schema, TypeSerializer<T> typeSerializer) {
 		this.tableSchema = SelectTableSinkSchemaConverter.convertTimeAttributeToRegularTimestamp(
-			SelectTableSinkSchemaConverter.changeDefaultConversionClass(schema));
+				SelectTableSinkSchemaConverter.changeDefaultConversionClass(schema));
 		this.converter = DataFormatConverters.getConverterForDataType(this.tableSchema.toPhysicalRowDataType());
 
 		String accumulatorName = "tableResultCollect_" + UUID.randomUUID();
@@ -80,7 +82,42 @@ public abstract class SelectTableSinkBase<T> implements StreamTableSink<T> {
 		return sink.name("Select table sink");
 	}
 
-	public abstract SelectResultProvider getSelectResultProvider();
+	public SelectResultProvider getSelectResultProvider() {
+		return new SelectResultProvider() {
+			@Override
+			public void setJobClient(JobClient jobClient) {
+				iterator.setJobClient(jobClient);
+			}
+
+			@Override
+			public Iterator<Row> getResultIterator() {
+				return new RowIteratorWrapper(iterator);
+			}
+		};
+	}
+
+	/**
+	 * An Iterator wrapper class that converts Iterator<T> to Iterator<Row>.
+	 */
+	private class RowIteratorWrapper implements Iterator<Row> {
+		private final CollectResultIterator<T> iterator;
+
+		public RowIteratorWrapper(CollectResultIterator<T> iterator) {
+			this.iterator = iterator;
+		}
+
+		@Override
+		public boolean hasNext() {
+			return iterator.hasNext();
+		}
+
+		@Override
+		public Row next() {
+			return convertToRow(iterator.next());
+		}
+	}
+
+	protected abstract Row convertToRow(T element);
 
 	/**
 	 * Create RowDataTypeInfo based on given table schema.
